+++ src/main/java/lotto/ui/InputView.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.ui;
 
 import java.util.Arrays;
 import java.util.List;
@@ -7,7 +7,7 @@ import java.util.stream.Collectors;
 
 public class InputView {
 
-    private Scanner scanner = new Scanner(System.in);
+    private final Scanner scanner = new Scanner(System.in);
 
     public int inputPurchaseAmount() {
         System.out.println("구입금액을 입력해 주세요.");
@@ -19,7 +19,7 @@ public class InputView {
     public List<Integer> inputWinningNumbers() {
         System.out.println("지난 주 당첨 번호를 입력해 주세요.");
         String inputs = scanner.nextLine();
-        return Arrays.stream(inputs.split(",")).map(input -> Integer.parseInt(input))
+        return Arrays.stream(inputs.split(",")).map(String::trim).map(input -> Integer.parseInt(input))
                 .collect(Collectors.toList());
     }
 }
